requirements.txt
glfw==2.10.0
numpy==2.3.4
pillow==12.0.0
pyglm==2.8.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
